refactor project and scrum compoenents Relates #247
